requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0
keras-tuner>=1.3.0

# Financial Analysis
yfinance>=0.2.0
ta-lib>=0.4.0
arch>=5.6.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
mplfinance>=0.12.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Logging & Monitoring
sentry-sdk>=1.28.0
